refactor: separate digest and macs
create own modules for utility functions/methods
