+++ kipher-digest/README.md
@@ -1,0 +1,3 @@
+# Module kipher-digest
+
+Library for cryptographic hash functions (such as SHAs, MD5s).
