+++ kipher-digest/build.gradle.kts
@@ -1,0 +1,10 @@
+plugins {
+    id("conventions.publish")
+}
+
+dependencies {
+    api(projects.kipherCommon)
+    api(projects.kipherUtils)
+    implementation(libs.bouncycastle.provider)
+    implementation(libs.jetbrains.annotations)
+}
