+++ kipher-digest/gradle.properties
@@ -1,0 +1,10 @@
+#
+# Copyright 2023 Joshua Hero Dela Cruz
+# SPDX-License-Identifier: Apache-2.0
+#
+
+ARTIFACT_ID=kipher-digest
+VERSION=0.1.0-SNAPSHOT
+
+POM_NAME=kipher-digest
+POM_DESCRIPTION=Library for cryptographic hash functions.
