--- kipher-hash/README.md
@@ -1,3 +1,0 @@
-# Module kipher-hash
-
-Library for ensuring data integrity and authenticity using cryptographic hash functions.
