--- kipher-hash/build.gradle.kts
@@ -1,9 +1,0 @@
-plugins {
-    id("conventions.publish")
-}
-
-dependencies {
-    api(projects.kipherCommon)
-    implementation(libs.bouncycastle.provider)
-    implementation(libs.jetbrains.annotations)
-}
