--- kipher-hash/gradle.properties
@@ -1,5 +1,0 @@
-ARTIFACT_ID=kipher-hash
-VERSION=0.1.0-SNAPSHOT
-
-POM_NAME=kipher-hash
-POM_DESCRIPTION=Library for ensuring data integrity and authenticity using cryptographic hash functions.
