--- kipher-hash/src/main/kotlin/io/github/jhdcruz/kipher/hash/digest/DigestSHA384.kt
@@ -1,11 +1,0 @@
-/*
- * Copyright 2023 Joshua Hero Dela Cruz
- * SPDX-License-Identifier: Apache-2.0
- */
-
-package io.github.jhdcruz.kipher.hash.digest
-
-/**
- * Cryptographic hashing using SHA-384.
- */
-class DigestSHA384 : Digest(DigestModes.SHA_384)
