--- kipher-hash/src/main/kotlin/io/github/jhdcruz/kipher/hash/digest/DigestSHA3_256.kt
@@ -1,8 +1,0 @@
-/*
- * Copyright 2023 Joshua Hero Dela Cruz
- * SPDX-License-Identifier: Apache-2.0
- */
-
-package io.github.jhdcruz.kipher.hash.digest
-
-class DigestSHA3_256 : Digest(DigestModes.SHA3_256)
