--- kipher-hash/src/main/kotlin/io/github/jhdcruz/kipher/hash/mac/HmacMD5.kt
@@ -1,11 +1,0 @@
-/*
- * Copyright 2023 Joshua Hero Dela Cruz
- * SPDX-License-Identifier: Apache-2.0
- */
-
-package io.github.jhdcruz.kipher.hash.mac
-
-/**
- * HMAC authentication using MD5.
- */
-class HmacMD5 : Hmac(MacModes.HmacMD5)
