--- kipher-hash/src/main/kotlin/io/github/jhdcruz/kipher/hash/mac/HmacSHA256.kt
@@ -1,11 +1,0 @@
-/*
- * Copyright 2023 Joshua Hero Dela Cruz
- * SPDX-License-Identifier: Apache-2.0
- */
-
-package io.github.jhdcruz.kipher.hash.mac
-
-/**
- * HMAC authentication using SHA-256.
- */
-class HmacSHA256 : Hmac(MacModes.HmacSHA256)
