--- kipher-hash/src/test/kotlin/io/github/jhdcruz/kipher/hash/digest/DigestModesTest.kt
@@ -1,132 +1,0 @@
-/*
- * Copyright 2023 Joshua Hero Dela Cruz
- * SPDX-License-Identifier: Apache-2.0
- */
-
-package io.github.jhdcruz.kipher.hash.digest
-
-import io.github.jhdcruz.kipher.hash.HashUtils.toHexString
-import kotlin.test.Test
-import kotlin.test.assertTrue
-
-internal class DigestModesTest {
-
-    private val data = "test".encodeToByteArray()
-
-    @Test
-    fun `test MD5 hash generation`() {
-        val digest = DigestMD5()
-
-        val hash = digest.generateHash(data)
-
-        println(hash.toHexString())
-
-        assertTrue {
-            digest.verifyHash(data, hash)
-        }
-    }
-
-    @Test
-    fun `test SHA-224 hash generation`() {
-        val digest = DigestSHA224()
-
-        val hash = digest.generateHash(data)
-
-        println(hash.toHexString())
-
-        assertTrue {
-            digest.verifyHash(data, hash)
-        }
-    }
-
-    @Test
-    fun `test SHA-256 hash generation`() {
-        val digest = DigestSHA256()
-
-        val hash = digest.generateHash(data)
-
-        println(hash.toHexString())
-
-        assertTrue {
-            digest.verifyHash(data, hash)
-        }
-    }
-
-    @Test
-    fun `test SHA-384 hash generation`() {
-        val digest = DigestSHA384()
-
-        val hash = digest.generateHash(data)
-
-        println(hash.toHexString())
-
-        assertTrue {
-            digest.verifyHash(data, hash)
-        }
-    }
-
-    @Test
-    fun `test SHA-512 hash generation`() {
-        val digest = DigestSHA512()
-
-        val hash = digest.generateHash(data)
-
-        println(hash.toHexString())
-
-        assertTrue {
-            digest.verifyHash(data, hash)
-        }
-    }
-
-    @Test
-    fun `test SHA3-224 hash generation`() {
-        val digest = DigestSHA3_224()
-
-        val hash = digest.generateHash(data)
-
-        println(hash.toHexString())
-
-        assertTrue {
-            digest.verifyHash(data, hash)
-        }
-    }
-
-    @Test
-    fun `test SHA3-256 hash generation`() {
-        val digest = DigestSHA3_256()
-
-        val hash = digest.generateHash(data)
-
-        println(hash.toHexString())
-
-        assertTrue {
-            digest.verifyHash(data, hash)
-        }
-    }
-
-    @Test
-    fun `test SHA3-384 hash generation`() {
-        val digest = DigestSHA3_384()
-
-        val hash = digest.generateHash(data)
-
-        println(hash.toHexString())
-
-        assertTrue {
-            digest.verifyHash(data, hash)
-        }
-    }
-
-    @Test
-    fun `test SHA3-512 hash generation`() {
-        val digest = DigestSHA3_512()
-
-        val hash = digest.generateHash(data)
-
-        println(hash.toHexString())
-
-        assertTrue {
-            digest.verifyHash(data, hash)
-        }
-    }
-}
