+++ kipher-mac/README.md
@@ -1,0 +1,3 @@
+# Module kipher-mac
+
+Library for ensuring data integrity and authenticity using message authentication codes.
