+++ kipher-mac/gradle.properties
@@ -1,0 +1,5 @@
+ARTIFACT_ID=kipher-mac
+VERSION=0.1.0-SNAPSHOT
+
+POM_NAME=kipher-mac
+POM_DESCRIPTION=Library for ensuring data integrity and authenticity using message authentication codes.
