+++ kipher-mac/src/main/kotlin/io/github/jhdcruz/kipher/mac/HmacMD5.kt
@@ -1,0 +1,11 @@
+/*
+ * Copyright 2023 Joshua Hero Dela Cruz
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+package io.github.jhdcruz.kipher.mac
+
+/**
+ * HMAC authentication using MD5.
+ */
+class HmacMD5 : Hmac(MacModes.HmacMD5)
