+++ kipher-mac/src/main/kotlin/io/github/jhdcruz/kipher/mac/HmacSHA256.kt
@@ -1,0 +1,12 @@
+/*
+ * Copyright 2023 Joshua Hero Dela Cruz
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+package io.github.jhdcruz.kipher.mac
+
+
+/**
+ * HMAC authentication using SHA-256.
+ */
+class HmacSHA256 : Hmac(MacModes.HmacSHA256)
