+++ kipher-mac/src/main/kotlin/io/github/jhdcruz/kipher/mac/HmacSHA384.kt
@@ -1,0 +1,11 @@
+/*
+ * Copyright 2023 Joshua Hero Dela Cruz
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+package io.github.jhdcruz.kipher.mac
+
+/**
+ * HMAC authentication using SHA-384.
+ */
+class HmacSHA384 : Hmac(MacModes.HmacSHA384)
