+++ kipher-utils/build.gradle.kts
@@ -1,0 +1,12 @@
+/*
+ * Copyright 2023 Joshua Hero Dela Cruz
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+plugins {
+    id("conventions.publish")
+}
+
+dependencies {
+    implementation(libs.jetbrains.annotations)
+}
